requirements.txt
# =============================================================================
# Sanjivani AI - Production Dependencies
# =============================================================================
# Install with: pip install -r requirements.txt

# -----------------------------------------------------------------------------
# Core Python
# -----------------------------------------------------------------------------
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# -----------------------------------------------------------------------------
# Machine Learning / Deep Learning
# -----------------------------------------------------------------------------
torch==2.1.2
torchvision==0.16.2
transformers==4.35.2
scikit-learn==1.3.2
xgboost==2.0.3
numpy==1.26.2
pandas==2.1.4
scipy==1.11.4

# -----------------------------------------------------------------------------
# Computer Vision
# -----------------------------------------------------------------------------
opencv-python==4.8.1.78
albumentations==1.3.1
segmentation-models-pytorch==0.3.3
ultralytics==8.0.227
Pillow==10.1.0

# -----------------------------------------------------------------------------
# Natural Language Processing
# -----------------------------------------------------------------------------
spacy==3.7.2
nltk==3.8.1
indic-transliteration==2.3.45
regex==2023.10.3

# -----------------------------------------------------------------------------
# Web Framework (FastAPI)
# -----------------------------------------------------------------------------
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
websockets==12.0

# -----------------------------------------------------------------------------
# Database
# -----------------------------------------------------------------------------
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
redis==5.0.1
alembic==1.13.0
geoalchemy2==0.14.2

# -----------------------------------------------------------------------------
# Dashboard (Streamlit)
# -----------------------------------------------------------------------------
streamlit==1.28.2
plotly==5.18.0
folium==0.15.1
streamlit-folium==0.15.1
altair==5.2.0

# -----------------------------------------------------------------------------
# Utilities
# -----------------------------------------------------------------------------
requests==2.31.0
aiofiles==23.2.1
python-dateutil==2.8.2
tqdm==4.66.1
pyyaml==6.0.1
orjson==3.9.10

# -----------------------------------------------------------------------------
# Logging and Monitoring
# -----------------------------------------------------------------------------
loguru==0.7.2
structlog==23.2.0

# -----------------------------------------------------------------------------
# Testing
# -----------------------------------------------------------------------------
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2
factory-boy==3.3.0
faker==21.0.0

# -----------------------------------------------------------------------------
# Code Quality
# -----------------------------------------------------------------------------
black==23.12.1
flake8==6.1.0
isort==5.13.2
mypy==1.7.1
pre-commit==3.6.0

# -----------------------------------------------------------------------------
# Jupyter Notebooks (Development)
# -----------------------------------------------------------------------------
jupyter==1.0.0
ipykernel==6.27.1
matplotlib==3.8.2
seaborn==0.13.0

# -----------------------------------------------------------------------------
# TensorFlow/Keras (for LSTM forecasting)
# -----------------------------------------------------------------------------
tensorflow==2.15.0
keras==2.15.0
